refactor: node spin up
